+++ NpcAdventure/Commander.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using Harmony;
 using NpcAdventure.StateMachine;
 using NpcAdventure.Story;
 using NpcAdventure.Utils;
@@ -26,6 +27,7 @@ namespace NpcAdventure
 
             consoleCommands.Add("npcadventure_eligible", "Make player eligible to recruit a companion (server or singleplayer only)", this.Eligible);
             consoleCommands.Add("npcadventure_recruit", "Recruit an NPC as companion (server or singleplayer only)", this.Recruit);
+            consoleCommands.Add("npcadventure_patches", "List harmony patches applied by NPC Adventures\n\nUsage: npcadventure_patches [recheck]\n\n- recheck - Recheck conflictiong patches", this.GetPatches);
             this.monitor.Log("Registered debug commands", LogLevel.Info);
         }
 
@@ -40,6 +42,27 @@ namespace NpcAdventure
             {
                 this.monitor.Log("Can't eligible player when game is not loaded, in non-adventure mode or not running on server!", LogLevel.Alert);
             }
+        }
+
+        private void GetPatches(string command, string[] args)
+        {
+            string describePatch(Patch p) => $"{p.patch.ReflectedType.FullName}.{p.patch.Name}";
+
+            if (args.Length > 0 && args[0] == "recheck")
+            {
+                this.npcAdventureMod.Patcher.CheckPatches(true);
+                return;
+            }
+
+            foreach (var patchedMethod in this.npcAdventureMod.Patcher.GetPatchedMethods())
+            {
+                this.monitor.Log($"{(patchedMethod.PatchInfo.Owners.Count > 1 ? "* " : "")}Patched method '{patchedMethod.Method.FullDescription()}'", LogLevel.Info);
+                this.monitor.Log($"   patched by: {string.Join(", ", patchedMethod.PatchInfo.Owners)}", LogLevel.Info);
+                this.monitor.Log($"   applied prefixes: {string.Join(", ", patchedMethod.PatchInfo.Prefixes.Select(describePatch))}", LogLevel.Info);
+                this.monitor.Log($"   applied postfixes: {string.Join(", ", patchedMethod.PatchInfo.Postfixes.Select(describePatch))}", LogLevel.Info);
+            }
+
+            this.monitor.Log("* - This patched method was patched by any other mod", LogLevel.Info);
         }
 
         private void Recruit(string command, string[] args)