+++ NpcAdventure/Internal/Patching/GamePatcher.cs
@@ -3,6 +3,7 @@ using StardewModdingAPI;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 
 namespace NpcAdventure.Internal.Patching
 {
@@ -11,30 +12,54 @@ namespace NpcAdventure.Internal.Patching
         private readonly IMonitor monitor;
         private readonly bool paranoid;
         private readonly HarmonyInstance harmony;
+        private readonly List<MethodBase> knownConflictedPatches;
 
         public GamePatcher(string uid, IMonitor monitor, bool paranoid = true)
         {
             this.monitor = monitor;
             this.paranoid = paranoid;
+            this.knownConflictedPatches = new List<MethodBase>();
             this.harmony = HarmonyInstance.Create(uid);
         }
 
-        public void CheckPatches()
+        /// <summary>
+        /// Returns all patched methods by NPC Adventures mod.
+        /// </summary>
+        /// <returns>Enumerable patched methods</returns>
+        public IEnumerable<PatchedMethod> GetPatchedMethods()
         {
+            var methods = this.harmony
+                .GetPatchedMethods()
+                .ToArray();
+            foreach (var method in methods)
+            {
+                Harmony.Patches info = this.harmony.GetPatchInfo(method);
+
+                if (info != null && info.Owners.Contains(this.harmony.Id))
+                {
+                    yield return new PatchedMethod(method, info);
+                }
+            }
+        }
+
+        public void CheckPatches(bool hard = false)
+        {
+            if (hard)
+            {
+                this.knownConflictedPatches.Clear();
+            }
+
             try
             {
-                var methods = this.harmony.GetPatchedMethods();
-
-                foreach (var method in methods)
+                foreach (var patchedMethod in this.GetPatchedMethods())
                 {
-                    Harmony.Patches info = this.harmony.GetPatchInfo(method);
-
-                    if (info.Owners.Contains(this.harmony.Id) && info.Owners.Count > 1)
+                    if (!this.knownConflictedPatches.Contains(patchedMethod.Method) && patchedMethod.PatchInfo.Owners.Count > 1)
                     {
-                        IEnumerable<string> foreignOwners = info.Owners.Where(owner => owner != this.harmony.Id);
+                        IEnumerable<string> foreignOwners = this.GetForeignOwners(patchedMethod);
 
-                        this.monitor.Log($"Detected another patches for game method '{method.FullDescription()}'. This method was patched too by: {string.Join(", ", foreignOwners)}",
+                        this.monitor.Log($"Detected another patches for game method '{patchedMethod.Method.FullDescription()}'. This method was patched too by: {string.Join(", ", foreignOwners)}",
                             this.paranoid ? LogLevel.Warn : LogLevel.Debug);
+                        this.knownConflictedPatches.Add(patchedMethod.Method);
                     }
                 }
             } catch (Exception ex)
@@ -42,6 +67,12 @@ namespace NpcAdventure.Internal.Patching
                 this.monitor.Log("Unable to check game patches. See log for more details.", LogLevel.Error);
                 this.monitor.Log(ex.ToString(), LogLevel.Trace);
             }
+        }
+
+        private IEnumerable<string> GetForeignOwners(PatchedMethod patchedMethod)
+        {
+            return patchedMethod.PatchInfo.Owners
+                .Where(owner => owner != this.harmony.Id);
         }
 
         /// <summary>
@@ -62,6 +93,18 @@ namespace NpcAdventure.Internal.Patching
                     this.monitor.Log(ex.ToString(), LogLevel.Trace);
                 }
             }
+        }
+
+        public class PatchedMethod
+        {
+            public PatchedMethod(MethodBase method, Harmony.Patches patchInfo)
+            {
+                this.Method = method;
+                this.PatchInfo = patchInfo;
+            }
+
+            public MethodBase Method { get; }
+            public Harmony.Patches PatchInfo { get; }
         }
     }
 }