+++ NpcAdventure/Model/Config.cs
@@ -22,7 +22,11 @@ namespace NpcAdventure.Model
 
         public class ExperimentalFeatures
         {
+            // From version 0.13.0 or 0.14.0 as stable option (enabled by default)
             public bool FightThruCompanion { get; set; } = false;
+            // From version 0.13.0 or 0.14.0 as stable option (enabled by default), 
+            // from version 0.16.0 will be removed and enabled hard
+            public bool UseCheckForEventsPatch { get; set; } = false;
         }
     }
 }